pyproject.toml
[project]
name = "tbl-rss"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "feedparser>=6.0.12",
  "httpx>=0.28.1",
  "openai>=2.7.1",
  "python-dotenv>=1.2.1",
]

[tool.basedpyright]
typeCheckingMode = 'standard'
useLibraryCodeForTypes = true
